Counting words
I wanted to a small script that would count the occurences of a hashtag
in a body of text. This small script was made in aproxximately 25-30
minutes, with a minor refactor at the end to keep it clean. All in all
it is super simple and gets the job done.